requirements.txt
pysrt==1.1.2
requests>=2.26.0,<3.0
pytest==8.3.4
pytest-mock==3.14.0
pytest-cov==6.0.0
openai==1.64.0
pydub==0.25.1
openai-whisper==20240930